move PerseusTester to conftest.py
